folder and file checking if present added

+++ downloadModpack.sh
@@ -4,13 +4,19 @@ set -euo pipefail
 if [ "$#" -ne 2 ]; then
 	if [ -z "$#" ]; then
 	echo -e "command is empty. please enter arguments"
-        exit 1
+	exit 1  
 	fi
 fi
-rm "$2"
+if [ -a "$2" ]; then
+        echo "deleting existing modpack install"
+        rm "$2"
+fi
 
-rm -r /opt/factorio/mods
+if [ -w /opt/factorio/mods ]; then
+        echo "deleting existing mods in factorio"
+        rm -r /opt/factorio/mods
+fi
 
-./gdrive download $1
+gdrive download $1
 
 7z x -o/opt/factorio/mods "$2"
